Move tests for apps into a subdirectory
