+++ nbgrader/tests/test_apps/test_nbgrader_validate.py
@@ -1,5 +1,9 @@
+import os
+
 from .base import TestBase
-from .utils import temp_cwd, run_command
+from ..utils import temp_cwd, run_command
+
+root = os.path.dirname(__file__)
 
 class TestNbgraderValidate(TestBase):
 
@@ -9,37 +13,37 @@ class TestNbgraderValidate(TestBase):
 
     def test_validate_unchanged(self):
         """Does the validation fail on an unchanged notebook?"""
-        with temp_cwd(["files/submitted-unchanged.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-unchanged.ipynb")]):
             output = run_command('nbgrader validate submitted-unchanged.ipynb')
             assert output.split("\n")[0] == "VALIDATION FAILED ON 2 CELL(S)! If you submit your assignment as it is, you WILL NOT"
 
     def test_validate_changed(self):
         """Does the validation pass on an changed notebook?"""
-        with temp_cwd(["files/submitted-changed.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-changed.ipynb")]):
             output = run_command('nbgrader validate submitted-changed.ipynb')
             assert output == "Success! Your notebook passes all the tests.\n"
 
     def test_invert_validate_unchanged(self):
         """Does the inverted validation pass on an unchanged notebook?"""
-        with temp_cwd(["files/submitted-unchanged.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-unchanged.ipynb")]):
             output = run_command('nbgrader validate submitted-unchanged.ipynb --invert')
             assert output.split("\n")[0] == "NOTEBOOK PASSED ON 1 CELL(S)!"
 
     def test_invert_validate_changed(self):
         """Does the inverted validation fail on a changed notebook?"""
-        with temp_cwd(["files/submitted-changed.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-changed.ipynb")]):
             output = run_command('nbgrader validate submitted-changed.ipynb --invert')
             assert output.split("\n")[0] == "NOTEBOOK PASSED ON 2 CELL(S)!"
 
     def test_grade_cell_changed(self):
         """Does the validate fail if a grade cell has changed?"""
-        with temp_cwd(["files/submitted-grade-cell-changed.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-grade-cell-changed.ipynb")]):
             output = run_command('nbgrader validate submitted-grade-cell-changed.ipynb')
             assert output.split("\n")[0] == "THE CONTENTS OF 1 TEST CELL(S) HAVE CHANGED! This might mean that even though the tests"
 
     def test_grade_cell_changed_ignore_checksums(self):
         """Does the validate pass if a grade cell has changed but we're ignoring checksums?"""
-        with temp_cwd(["files/submitted-grade-cell-changed.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-grade-cell-changed.ipynb")]):
             output = run_command(
                 'nbgrader validate submitted-grade-cell-changed.ipynb '
                 '--DisplayAutoGrades.ignore_checksums=True')
@@ -47,13 +51,13 @@ class TestNbgraderValidate(TestBase):
 
     def test_invert_grade_cell_changed(self):
         """Does the validate fail if a grade cell has changed, even with --invert?"""
-        with temp_cwd(["files/submitted-grade-cell-changed.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-grade-cell-changed.ipynb")]):
             output = run_command('nbgrader validate submitted-grade-cell-changed.ipynb --invert')
             assert output.split("\n")[0] == "THE CONTENTS OF 1 TEST CELL(S) HAVE CHANGED! This might mean that even though the tests"
 
     def test_invert_grade_cell_changed_ignore_checksums(self):
         """Does the validate fail if a grade cell has changed with --invert and ignoring checksums?"""
-        with temp_cwd(["files/submitted-grade-cell-changed.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-grade-cell-changed.ipynb")]):
             output = run_command(
                 'nbgrader validate submitted-grade-cell-changed.ipynb '
                 '--invert '
@@ -62,7 +66,7 @@ class TestNbgraderValidate(TestBase):
 
     def test_validate_unchanged_ignore_checksums(self):
         """Does the validation fail on an unchanged notebook with ignoring checksums?"""
-        with temp_cwd(["files/submitted-unchanged.ipynb"]):
+        with temp_cwd([os.path.join(root, "files/submitted-unchanged.ipynb")]):
             output = run_command(
                 'nbgrader validate submitted-unchanged.ipynb '
                 '--DisplayAutoGrades.ignore_checksums=True')
